refactor: move client-related identifiers to id.ts and remove shared client module
