+++ packages/chapplin/src/vite/id.ts
@@ -1,0 +1,4 @@
+export const id = "/__virtual-chapplin";
+export const idRegex = new RegExp(`^${id}$`);
+export const resolvedId = `\0__virtual-chapplin`;
+export const resolvedIdRegex = new RegExp(`^${resolvedId}$`);
