+++ packages/chapplin/src/vite/plugins/client-tool-resolver.ts
@@ -1,5 +1,5 @@
+import type { PluginContext } from "rolldown";
 import type { Plugin, ResolvedConfig } from "vite";
-import { resolveTarget } from "../shared/client.js";
 import type { Options, Target } from "../types.js";
 
 export function clientToolResolver(opts?: Options): Plugin {
@@ -28,4 +28,58 @@ export function clientToolResolver(opts?: Options): Plugin {
 			},
 		},
 	};
+}
+
+function getJsxImportSourceFromResolvedConfig(
+	config: ResolvedConfig,
+): string | undefined {
+	const jsx = config.build?.rollupOptions?.jsx;
+	if (typeof jsx !== "object") return;
+	return jsx.importSource ?? undefined;
+}
+
+async function getJsxImportSourceFromTsconfig(
+	fs: PluginContext["fs"],
+	tsconfigPath: string,
+): Promise<string | undefined> {
+	try {
+		// tsconfig allows comments and trailing commas,
+		// so we need to do some manual parsing here
+		const tsconfig = await fs.readFile(tsconfigPath, { encoding: "utf8" });
+		const tsconfigLines = tsconfig
+			.split("\n")
+			.map((ln) => ln.trim())
+			.filter((ln) => !ln.startsWith("//"));
+		const jsxImportSource = tsconfigLines
+			.find((ln) => ln.startsWith('"jsxImportSource"'))
+			?.match(/"jsxImportSource"[ \t]*:[ \t]*"(.+?)"/)?.[1];
+		return jsxImportSource;
+	} catch {
+		// noop
+	}
+}
+
+function resolveTargetFromJsxImportSource(jsxImportSource: string): Target {
+	if (jsxImportSource === "react") return "react";
+	if (jsxImportSource === "preact") return "preact";
+	if (jsxImportSource.startsWith("hono/jsx")) return "hono";
+	return "react"; // Default to react
+}
+
+async function resolveTarget(
+	resolvedConfig: ResolvedConfig,
+	fs: PluginContext["fs"],
+	opts?: Pick<Options, "target" | "tsconfigPath">,
+): Promise<Target> {
+	if (opts?.target) return opts.target;
+
+	const jsxImportSource =
+		getJsxImportSourceFromResolvedConfig(resolvedConfig) ??
+		(await getJsxImportSourceFromTsconfig(
+			fs,
+			opts?.tsconfigPath ?? "tsconfig.json",
+		)) ??
+		"react";
+
+	return resolveTargetFromJsxImportSource(jsxImportSource);
 }
